started work on refactoring entry ad tests
